+++ ios/ios/src/Storage/CMStore.m
@@ -348,13 +348,13 @@ NSString * const CMStoreObjectDeletedNotification = @"CMStoreObjectDeletedNotifi
                             user:_CMUserOrNil
                  extraParameters:_CMTryMethod(options, buildExtraParameters)
                   successHandler:^(NSDictionary *results, NSDictionary *errors, NSDictionary *meta, NSDictionary *snippetResult, NSNumber *count, NSDictionary *headers) {
-                      NSLog(@"ERROR: %@", errors);
+
                       NSArray *objects = [CMObjectDecoder decodeObjects:results];
                       [self cacheObjectsInMemory:objects atUserLevel:userLevel];
                       CMResponseMetadata *metadata = [[CMResponseMetadata alloc] initWithMetadata:meta];
                       CMSnippetResult *result = [[CMSnippetResult alloc] initWithData:snippetResult];
                       CMObjectFetchResponse *response = [[CMObjectFetchResponse alloc] initWithObjects:objects errors:errors snippetResult:result responseMetadata:metadata];
-                      response.count = count ? [count intValue] : [objects count];
+                      response.count = count ? [count integerValue] : [objects count];
 
                       [objects enumerateObjectsUsingBlock:^(CMObject *obj, NSUInteger idx, BOOL *stop) {
                           obj.ownerId = [metadata metadataForObject:obj ofType:@"owner"];